@@ -1,47 +1,16 @@
 {
   "name": "bloomreach-content-saas-ui-extensions",
-  "version": "1.4.0",
+  "version": "2.0.0",
   "title": "Bloomreach Content SaaS UI Extensions",
   "description": "High performance React template built with lots of powerful MUI (Material-UI) components across multiple product niches for fast & perfect apps development processes",
   "author": {
     "name": "Bloomreach",
     "url": "https://www.bloomreach.com"
   },
-  "private": false,
-  "dependencies": {
-    "@emotion/cache": "11.7.1",
-    "@emotion/react": "11.9.0",
-    "@emotion/server": "11.4.0",
-    "@emotion/styled": "11.8.1",
-    "@mui/core": "5.0.0-alpha.54",
-    "@mui/icons-material": "5.8.2",
-    "@mui/lab": "5.0.0-alpha.84",
-    "@mui/material": "5.8.2",
-    "@mui/styles": "5.8.0",
-    "@mui/x-data-grid": "^5.17.20",
-    "@reduxjs/toolkit": "^1.9.2",
-    "apexcharts": "3.35.3",
-    "axios": "^1.2.2",
-    "bloomreach-content-management-apis": "^1.2.0",
-    "clsx": "1.1.1",
-    "context-api": "0.0.2",
-    "date-fns": "2.28.0",
-    "dependency-graph": "^0.11.0",
-    "gray-matter": "^4.0.3",
-    "next": "^13.4.4",
-    "next-images": "1.8.4",
-    "nprogress": "0.2.0",
-    "numeral": "2.0.6",
-    "react": "^18.2.0",
-    "react-apexcharts": "1.4.0",
-    "react-code-blocks": "*",
-    "react-custom-scrollbars-2": "4.4.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^8.0.5",
-    "remark": "^14.0.2",
-    "remark-gfm": "^3.0.1",
-    "remark-html": "^15.0.2"
+  "engines": {
+    "node": ">=20.0.0"
   },
+  "private": false,
   "scripts": {
     "dev": "next",
     "build": "next build",
@@ -55,18 +24,57 @@
     "test:watch": "node -r dotenv/config $(which jest) --watchAll",
     "yalcAdd": "yalc add bloomreach-content-management-apis"
   },
+  "dependencies": {
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "11.9.0",
+    "@emotion/server": "11.4.0",
+    "@emotion/styled": "11.8.1",
+    "@mui/core": "5.0.0-alpha.54",
+    "@mui/icons-material": "5.8.2",
+    "@mui/lab": "5.0.0-alpha.84",
+    "@mui/material": "5.8.2",
+    "@mui/material-nextjs": "^5.15.11",
+    "@mui/styled-engine": "^5.15.14",
+    "@mui/styles": "5.8.0",
+    "@mui/x-data-grid": "^5.17.20",
+    "@mui/x-tree-view": "^7.1.1",
+    "axios": "^1.2.2",
+    "bloomreach-content-management-apis": "file:.yalc/bloomreach-content-management-apis",
+    "clsx": "1.1.1",
+    "context-api": "0.0.2",
+    "date-fns": "2.28.0",
+    "dependency-graph": "^0.11.0",
+    "eslint-config-next": "^14.2.3",
+    "gray-matter": "^4.0.3",
+    "next": "^14.2.4",
+    "next-images": "1.8.4",
+    "nprogress": "0.2.0",
+    "numeral": "2.0.6",
+    "react": "^18.2.0",
+    "react-apexcharts": "1.4.0",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-code-blocks": "*",
+    "react-custom-scrollbars-2": "4.4.0",
+    "react-dom": "^18.2.0",
+    "remark": "^14.0.2",
+    "remark-gfm": "^3.0.1",
+    "remark-html": "^15.0.2"
+  },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
     "dotenv": "^16.0.3",
     "eslint": "8.17.0",
-    "eslint-config-next": "^13.4.4",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-unused-imports": "^3.1.0",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
-    "next-transpile-modules": "9.0.0",
+    "next-transpile-modules": "^10.0.1",
     "prettier": "2.6.2",
     "yalc": "1.0.0-pre.53"
+  },
+  "resolutions": {
+    "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest"
   }
 }
