@@ -1,7 +1,8 @@
+'use client'
 import * as React from 'react';
-import PropTypes from 'prop-types';
 import clsx from 'clsx';
-import { useRouter } from 'next/router';
+// import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation'
 import NextLink from 'next/link';
 import MuiLink from '@mui/material/Link';
 import { styled } from '@mui/material/styles';
@@ -32,18 +33,6 @@ export const NextLinkComposed = React.forwardRef(function NextLinkComposed(
   );
 });
 
-NextLinkComposed.propTypes = {
-  href: PropTypes.any,
-  linkAs: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
-  locale: PropTypes.string,
-  passHref: PropTypes.bool,
-  prefetch: PropTypes.bool,
-  replace: PropTypes.bool,
-  scroll: PropTypes.bool,
-  shallow: PropTypes.bool,
-  to: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired
-};
-
 // A styled version of the Next.js Link component:
 // https://nextjs.org/docs/api-reference/next/link
 const Link = React.forwardRef(function Link(props, ref) {
@@ -57,7 +46,6 @@ const Link = React.forwardRef(function Link(props, ref) {
     noLinkStyle,
     prefetch,
     replace,
-    role, // Link don't have roles.
     scroll,
     shallow,
     ...other
@@ -113,20 +101,5 @@ const Link = React.forwardRef(function Link(props, ref) {
     />
   );
 });
-
-Link.propTypes = {
-  activeClassName: PropTypes.string,
-  as: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
-  className: PropTypes.string,
-  href: PropTypes.any,
-  linkAs: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
-  locale: PropTypes.string,
-  noLinkStyle: PropTypes.bool,
-  prefetch: PropTypes.bool,
-  replace: PropTypes.bool,
-  role: PropTypes.string,
-  scroll: PropTypes.bool,
-  shallow: PropTypes.bool
-};
 
 export default Link;
