@@ -1,38 +1,38 @@
 const bloomreachTheme = {
-  lineNumberColor: `white`,
-  lineNumberBgColor: `#002840`,
-  backgroundColor: `#002840`,
-  textColor: `#d8d8d8`,
-  substringColor: `#d8d8d8`,
-  keywordColor: `#ebbbff`,
-  attributeColor: `#ffd500`,
-  selectorAttributeColor: `#ebbbff`,
-  docTagColor: `white`,
-  nameColor: `#ff9da4`,
-  builtInColor: `#ffc58f`,
-  literalColor: `#ffc58f`,
-  bulletColor: `#ffd500`,
-  codeColor: `white`,
-  additionColor: `#ffd500`,
-  regexpColor: `#ff9da4`,
-  symbolColor: `#ffd500`,
-  variableColor: `#ff9da4`,
-  templateVariableColor: `#ff9da4`,
-  linkColor: `#ffc58f`,
-  selectorClassColor: `#ff9da4`,
-  typeColor: `#ffc58f`,
-  stringColor: `#ffd500`,
-  selectorIdColor: `#ff9da4`,
-  quoteColor: `#7285b7`,
-  templateTagColor: `#d8d8d8`,
-  deletionColor: `#ff9da4`,
-  titleColor: `#bbdaff`,
-  sectionColor: `#00b2db`,
-  commentColor: `#7285b7`,
-  metaKeywordColor: `white`,
-  metaColor: `#ffc58f`,
-  functionColor: `white`,
-  numberColor: `#ffc58f`,
+  lineNumberColor: 'white',
+  lineNumberBgColor: '#002840',
+  backgroundColor: '#002840',
+  textColor: '#d8d8d8',
+  substringColor: '#d8d8d8',
+  keywordColor: '#ebbbff',
+  attributeColor: '#ffd500',
+  selectorAttributeColor: '#ebbbff',
+  docTagColor: 'white',
+  nameColor: '#ff9da4',
+  builtInColor: '#ffc58f',
+  literalColor: '#ffc58f',
+  bulletColor: '#ffd500',
+  codeColor: 'white',
+  additionColor: '#ffd500',
+  regexpColor: '#ff9da4',
+  symbolColor: '#ffd500',
+  variableColor: '#ff9da4',
+  templateVariableColor: '#ff9da4',
+  linkColor: '#ffc58f',
+  selectorClassColor: '#ff9da4',
+  typeColor: '#ffc58f',
+  stringColor: '#ffd500',
+  selectorIdColor: '#ff9da4',
+  quoteColor: '#7285b7',
+  templateTagColor: '#d8d8d8',
+  deletionColor: '#ff9da4',
+  titleColor: '#bbdaff',
+  sectionColor: '#00b2db',
+  commentColor: '#7285b7',
+  metaKeywordColor: 'white',
+  metaColor: '#ffc58f',
+  functionColor: 'white',
+  numberColor: '#ffc58f',
 }
 
 export default bloomreachTheme
